cleaned up database logic and & added .env support
