+++ .env
@@ -1,0 +1,1 @@
+DB=mongodb://root:root@127.0.0.1:27017/
