+++ main.go
@@ -2,11 +2,17 @@ package main
 
 import (
 	"github.com/h0rzn/monitoring_agent/api"
+	"github.com/joho/godotenv"
 	"github.com/sirupsen/logrus"
 )
 
 func main() {
 	logrus.SetLevel(logrus.DebugLevel)
+	err := godotenv.Load(".env")
+	if err != nil {
+		logrus.Errorf("- MAIN - failed to load .env")
+		return
+	}
 	logrus.Infoln("starting metawach-agent")
 	api, err := api.NewAPI(":8080")
 	if err != nil {
